_backup/taebin-ai-agent/ai-drkube/requirements.txt
# LangGraph 및 LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0

# Kubernetes 클라이언트
kubernetes>=28.0.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 데이터 검증
pydantic>=2.0.0

# YAML 파싱
pyyaml>=6.0

# 로깅
loguru>=0.7.0

# 비동기 HTTP (선택사항)
aiohttp>=3.9.0

# 웹 서버 (Alertmanager Webhook Receiver용)
flask>=3.0.0
flask-cors>=4.0.0  # CORS 지원

# 1. **로그 수집**: 로컬 파일 또는 외부 소스에서 로그 수집
# 2. **에러 분류**: 키워드 기반 에러 카테고리 분류
# 3. **근본 원인 분석**: 에러 카테고리 기반 원인 분석
# 4. **사용자 피드백**: 분석 결과를 사용자에게 표시 (Phase 1: 자동 승인)
# 5. **액션 생성**: Git 변경사항 생성
# 6. **Git 커밋**: 변경사항 커밋 (Phase 1: 시뮬레이션)
# 7. **복구 검증**: 복구 결과 확인 (Phase 1: 시뮬레이션)
# 8. **최종 피드백**: 사용자에게 최종 결과 보고
# 1. **로그 수집**: 로컬 파일 또는 외부 소스에서 로그 수집
# 2. **에러 분류**: 키워드 기반 에러 카테고리 분류
# 3. **근본 원인 분석**: 에러 카테고리 기반 원인 분석
# 4. **사용자 피드백**: 분석 결과를 사용자에게 표시 (Phase 1: 자동 승인)
# 5. **액션 생성**: Git 변경사항 생성
# 6. **Git 커밋**: 변경사항 커밋 (Phase 1: 시뮬레이션)
# 7. **복구 검증**: 복구 결과 확인 (Phase 1: 시뮬레이션)
# 8. **최종 피드백**: 사용자에게 최종 결과 보고
